g++ compile warning fixes

+++ GLG3D.lib/source/Shader.cpp
@@ -208,8 +208,8 @@ void VertexAndPixelShader::GPUShader::replaceG3DSize(std::string& code, std::str
 
         size_t last = 0;
         while (true) {
-            int next = code.find(funcName, last);
-            if ((unsigned int)next == std::string::npos) {
+            size_t next = code.find(funcName, last);
+            if (next == std::string::npos) {
                 break;
             }
